+++ packages/x-adapter-platform/src/mappers/responses/experience-controls-response.mapper.ts
@@ -1,0 +1,15 @@
+import { schemaMapperFactory } from '@empathyco/x-adapter';
+import { ExperienceControlsResponse } from '@empathyco/x-types';
+import { PlatformExperienceControlsResponse } from '../../types';
+// eslint-disable-next-line max-len
+import { experienceControlsResponseSchema } from '../../schemas/responses/experience-controls-response.schema';
+
+/**
+ * Default implementation for the ExperienceControlsResponseMapper.
+ *
+ * @public
+ */
+export const experienceControlsResponseMapper = schemaMapperFactory<
+  PlatformExperienceControlsResponse,
+  ExperienceControlsResponse
+>(experienceControlsResponseSchema);