+++ packages/x-components/src/__tests__/utils.ts
@@ -12,7 +12,8 @@ import {
   QuerySuggestionsResponse,
   RelatedTagsResponse,
   SearchResponse,
-  SemanticQueriesResponse
+  SemanticQueriesResponse,
+  ExperienceControlsResponse
 } from '@empathyco/x-types';
 import { XPluginOptions } from '../plugins';
 import { XPlugin } from '../plugins/x-plugin';
@@ -45,6 +46,7 @@ interface MockedAdapterFeatures {
   search: SearchResponse;
   semanticQueries: SemanticQueriesResponse;
   tagging: void;
+  experienceControls: ExperienceControlsResponse;
 }
 
 /**
@@ -158,7 +160,8 @@ export function getMockedAdapter(
     relatedTags: getMockedAdapterFunction(responseFeatures?.relatedTags!),
     search: getMockedAdapterFunction(responseFeatures?.search!),
     semanticQueries: getMockedAdapterFunction(responseFeatures?.semanticQueries!),
-    tagging: getMockedAdapterFunction(undefined)
+    tagging: getMockedAdapterFunction(undefined),
+    experienceControls: getMockedAdapterFunction(responseFeatures?.experienceControls!)
     /* eslint-enable @typescript-eslint/no-non-null-asserted-optional-chain */
   };
 }