+++ packages/x-components/src/x-modules/experience-controls/store/emitters.ts
@@ -1,0 +1,14 @@
+import { createStoreEmitters } from '../../../store';
+import { experienceControlsXStoreModule } from './module';
+
+/**
+ * {@link StoreEmitters} For the {@link ExperienceControlsXModule}.
+ *
+ * @internal
+ */
+export const experienceControlsEmitters = createStoreEmitters(experienceControlsXStoreModule, {
+  ExperienceControlsEventsChanged: {
+    selector: state => state.events
+  },
+  ExperienceControlsRequestUpdated: (_, getters) => getters.experienceControlsRequest
+});